requirements.txt
anvil-parser==0.6.4
astroid==2.4.2
autopep8==1.5.3
frozendict==1.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
NBT==1.5.0
pycodestyle==2.6.0
pylint==2.5.3
rope==0.17.0
six==1.15.0
toml==0.10.1
tqdm==4.47.0
wrapt==1.12.1
